pyproject.toml
[project]
name = "yahboom-mcp"
version = "2.3.0"
description = "SOTA 2026 Yahboom Raspbot v2 ROS 2 MCP Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp==3.2.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "roslibpy>=1.2.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.23.0",
    "structlog>=23.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "opencv-python>=4.8.0",
    "python-miio>=0.5.12",
    "paramiko>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-timeout>=2.3.0",
    "httpx[asyncio]>=0.27.0",   # ASGI test client for snapshot endpoint tests
    "anyio>=4.0.0",              # required by httpx ASGI transport
    "ruff>=0.9.0",
]
# Same packages the MCP installs on the Pi over SSH for voice + OLED; optional locally for smoke tests.
robot-pi = [
    "pyserial>=3.5",
    "luma.oled>=3.12.0",
    "luma.core>=2.4.0",
    "pillow>=10.0.0",
    "smbus2>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/yahboom_mcp"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "unit: fast, no I/O",
    "mock: uses mock bridge / fakes (default CI)",
    "integration: ASGI, HTTP, or multi-module",
    "e2e: full hardware end-to-end (needs YAHBOOM_E2E=1 and physical robot)",
    "needs_robot: alias for e2e",
    "real_robot: alias for e2e",
    "slow: optional long-running",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
branch = true
source = ["yahboom_mcp"]
omit = ["*/testing/*", "*/tests/*"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "S", "UP", "RUF"]
ignore = ["S101", "B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
